hw10

+++ src/main/java/ua/orlov/hw10/Lambda.java
@@ -10,18 +10,18 @@ package ua.orlov.hw10;
 public class Lambda {
     public static void main(String[] args) {
         Pow op = (number, pow) -> {
-            int numstep = 1;
+            double numstep = 1;
             for (int i = 1; i <= pow; i++) {
                 numstep *= number;
             }
             return numstep;
         };
-        int result = op.pow(2, 3);
+        double result = op.pow(2, 3);
         System.out.println(result);
     }
 }
 
 interface Pow {
-    int pow(int number, int pow);
+    double pow(double number, int pow);
 }
